Add delete methods, split product adding and updating
Separate setting products and params of custom fields
from the main update method that effectively sends the update
request to searchanise. ``update`` now works kinda like a
``commit`` method, that can also optionally accept a list
of products and a dict of params as a convenience method
to do all from a single method call.